🏗 vue-dot: Reorganize tests
