+++ packages/vue-dot/src/functions/copyToClipboard/index.ts
