+++ packages/vue-dot/src/functions/getFileExtension/index.ts
