+++ packages/vue-dot/src/functions/getFileExtension/tests/getFileExtension.spec.ts
@@ -1,0 +1,10 @@
+import getFileExtension from '../';
+
+// Tests
+describe('getFileExtension', () => {
+	it('returns the extension of a file', () => {
+		const filename = 'test.pdf';
+
+		expect(getFileExtension(filename)).toEqual('pdf');
+	});
+});
