--- packages/vue-dot/src/rules/required.ts
@@ -1,31 +1,0 @@
-import ruleMessage from '../helpers/ruleMessage';
-import { ErrorMessages } from './types';
-
-const defaultErrorMessages: ErrorMessages = {
-	default: 'Le champ est requis'
-};
-
-/** Check that the field is non-empty */
-export function required(errorMessages = defaultErrorMessages) {
-	// The value can be an array of string in select with the multiple prop
-	return (value: string | string[]) => {
-		let valid: boolean;
-
-		// If the value is an array
-		if (Array.isArray(value)) {
-			// Check it's length
-			valid = value.length !== 0;
-		} else {
-			// If the value is falsy (empty string),
-			// it will return the error message
-
-			// If the value is evaluated to true (non-empty string),
-			// it will return true (success)
-			valid = Boolean(value);
-		}
-
-		return valid || ruleMessage(errorMessages, 'default');
-	};
-}
-
-export default required();
