--- packages/vue-dot/tests/functions/getFileExtension.spec.ts
@@ -1,10 +1,0 @@
-import getFileExtension from '../../src/functions/getFileExtension';
-
-// Tests
-describe('getFileExtension', () => {
-	it('returns the extension of a file name', () => {
-		const filename = 'test.pdf';
-
-		expect(getFileExtension(filename)).toEqual('pdf');
-	});
-});
